+++ R/readAndMergeEuCodedFiles.R
@@ -31,14 +31,18 @@ readAndMergeEuCodedFiles <- function(
   error.file = NULL
 )
 {
+  dot.args <- list(...)
+  #dot.args <- list() # for debugging!
+  
   # by setting simple.algorithm = FALSE we get unique column names, e.g. "ADE"
   # and "ADE.1"
-  inspection.data.list <- readEuCodedFiles(
+  inspection.data.list <- kwb.utils::callWith(
+    readEuCodedFiles,
     input.files = input.files, 
     dbg = dbg, 
     name.convention = name.convention,
     simple.algorithm = FALSE, 
-    ...
+    dot.args
   )
   
   result <- mergeInspectionData(inspection.data.list)