+++ R/readEuCodedFiles.R
@@ -33,10 +33,15 @@ readEuCodedFiles <- function(
       "input file %d/%d: %s\n", i, length(input.files), input.file
     ))
     
-    inspectionData <- try(
-      readEuCodedFile(input.file, dbg = dbg, ...),
-      silent = TRUE
-    )
+    dot.args <- list(...)
+    #dot.args <- list() # for debugging!
+    
+    inspectionData <- try(silent = TRUE, kwb.utils::callWith(
+      readEuCodedFile,
+      input.file = input.file, 
+      dbg = dbg, 
+      dot.args
+    ))
     
     # Return the error object if an error occurred
     if (kwb.utils::isTryError(inspectionData)) {