+++ tests/testthat/test-function-extractInspectionData_v2.R
@@ -1,0 +1,6 @@
+test_that("extractInspectionData_v2() works", {
+
+  f <- kwb.en13508.2:::extractInspectionData_v2
+  
+  expect_error(f())
+})