+++ tests/testthat/test-function-getInspectionRecordsFromEuLines.R
@@ -1,7 +1,0 @@
-test_that("getInspectionRecordsFromEuLines() works", {
-
-  f <- kwb.en13508.2:::getInspectionRecordsFromEuLines
-
-  expect_error(f())
-
-})